requirement.txt
fastapi
numpy
Pillow
python-version
torch
torchvision
typing_extensions
uvicorn
python-multipart
matplotlib
